+++ aql-engine/src/main/java/org/ehrbase/openehr/aqlengine/pathanalysis/PathInfo.java
@@ -236,7 +236,9 @@ public final class PathInfo {
                                 .map(p -> Pair.of(p, QueryClause.ORDER_BY)))
                 .flatMap(s -> s)
                 .collect(Collectors.groupingBy(
-                        Pair::getLeft, Collectors.mapping(Pair::getRight, Collectors.toUnmodifiableSet()))));
+                        Pair::getLeft,
+                        LinkedHashMap::new,
+                        Collectors.mapping(Pair::getRight, Collectors.toUnmodifiableSet()))));
 
         return containsDescs.entrySet().stream()
                 .filter(e -> pathCohesion.containsKey(e.getKey()))