+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0]
+ ### Added
+- Configurable flyway migration strategy
+- Configurable fetch limit checks + default limit for AQL queries
+- Configurable fetch limit precedence strategy for AQL queries
+ ### Changed 
+ - Upgrade openEHR_SDK to version 2.13.0 see https://github.com/ehrbase/openEHR_SDK/blob/develop/CHANGELOG.md 
+ ### Fixed 
+
 ## [2.3.0]
  ### Added
  ### Changed 
@@ -42,3 +51,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.1.0]: https://github.com/ehrbase/ehrbase/compare/v2.0.0...v2.1.0
 [2.2.0]: https://github.com/ehrbase/ehrbase/compare/v2.1.0...v2.2.0
 [2.3.0]: https://github.com/ehrbase/ehrbase/compare/v2.2.0...v2.3.0
+[2.4.0]: https://github.com/ehrbase/ehrbase/compare/v2.3.0...v2.4.0